rag-service/requirements.txt
fastapi>=0.95.2
uvicorn>=0.22.0
llama-index>=0.8.0
llama-index-vector-stores-chroma>=0.1.0
chromadb>=0.4.0
google-generativeai>=0.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0
pydantic-settings
llama-index-llms-gemini
llama-index-embeddings-gemini
openpyxl>=3.0.0
python-docx
docx2txt
pandas
